git-tfs-id: [http://192.168.56.101:8080/tfs/DefaultCollection]$/TikiEngine 2.0;C815

+++ src/Graphics/SpriteBatchModule.cpp
@@ -146,7 +146,11 @@ namespace TikiEngine
 		void SpriteBatchModule::Draw(ITexture* texture, const Rectangle& destRect)
 		{
 			Vector3 tl = transformPoint(
-				Vector3(destRect.X, destRect.Y, 0.0f)
+				Vector3(
+					(float)destRect.X,
+					(float)destRect.Y,
+					0.0f
+				)
 			); 
 
 			Vector3 br = Vector3(
@@ -168,7 +172,11 @@ namespace TikiEngine
 		void SpriteBatchModule::Draw(ITexture* texture, const Rectangle& destRect, const Rectangle& srcRect)
 		{
 			Vector3 tl = transformPoint(
-				Vector3(destRect.X, destRect.Y, 0.0f)
+				Vector3(
+					(float)destRect.X,
+					(float)destRect.Y,
+					0.0f
+				)
 			); 
 
 			Vector3 br = Vector3(
